v0.4

+++ app/router/blog/index.js
@@ -30,6 +30,7 @@ router.get('/dashboard/blogs', (req, res) => {
             userinfo = res2[0]
             if (userinfo == null) { return res.redirect('/auth/login') }
             if (userinfo.remember_token !== req.signedCookies.remember_token) { return res.redirect('/auth/login') } else {
+            if (userinfo.group != 'manager') { return res.redirect('/dashboard') }
                 connection.query('SELECT * FROM blogs', function(err, res2) {
                     return res.render('./dashboard/blog/all', {userinfo: userinfo, blogs: res2})
                 });
@@ -47,6 +48,7 @@ router.get('/dashboard/blogs/edit', (req, res) => {
                 userinfo = res2[0]
                 if (userinfo == null) { return res.redirect('/auth/login') }
                 if (userinfo.remember_token !== req.signedCookies.remember_token) { return res.redirect('/auth/login') } else {
+                if (userinfo.group != 'manager') { return res.redirect('/dashboard') }
                     connection.query('SELECT * FROM blogs WHERE id = ?', [req.query.id], function(err, res2) {
                         return res.render('./dashboard/blog/edit', {userinfo: userinfo, blogs: res2[0]})
                     });
@@ -66,6 +68,7 @@ router.post('/dashboard/blogs/edit', function (req, res) {
             userinfo = res2[0]
             if (userinfo == null) { return res.redirect('/auth/login') }
             if (userinfo.remember_token !== req.signedCookies.remember_token) { return res.redirect('/auth/login') } else {
+            if (userinfo.group != 'manager') { return res.redirect('/dashboard') }
                 if (!req.query.new) {
                     if (req.query.id) {
                         connection.query(`UPDATE blogs SET title = '${req.body.title}', about = '${req.body.about}', picture = '${req.body.picture}', id = '${req.body.id}', month = '${req.body.month}', day = '${req.body.day}', bigtext = '${req.body.bigtext}' WHERE id = '${req.query.id}'`)
@@ -103,6 +106,7 @@ router.get('/dashboard/blogs/new', (req, res) => {
             userinfo = res2[0]
             if (userinfo == null) { return res.redirect('/auth/login') }
             if (userinfo.remember_token !== req.signedCookies.remember_token) { return res.redirect('/auth/login') } else {
+            if (userinfo.group != 'manager') { return res.redirect('/dashboard') }
                     return res.render('./dashboard/blog/new', {userinfo: userinfo})
             }
         });
@@ -118,6 +122,7 @@ router.post('/dashboard/blogs/delete', function (req, res) {
                 userinfo = res2[0]
                 if (userinfo == null) { return res.redirect('/auth/login') }
                 if (userinfo.remember_token !== req.signedCookies.remember_token) { return res.redirect('/auth/login') } else {
+                if (userinfo.group != 'manager') { return res.redirect('/dashboard') }
                     if (req.query.id) {
                         connection.query("DELETE FROM blogs WHERE id = ?;", [req.query.id])
                             return res.redirect('/dashboard/blogs')
